# docs/openapi/b2e-agent.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "B2E agent",
  "description": "The agent under test. Multi-tenant, fully traced.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/sessions": {
   "post": {
    "summary": "Create Session",
    "operationId": "create_session_sessions_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateSession"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Session Sessions Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "summary": "List Sessions",
    "operationId": "list_sessions_sessions_get",
    "parameters": [
     {
      "name": "employee_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Employee Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Sessions Sessions Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/experiments": {
   "post": {
    "summary": "Create Experiment",
    "operationId": "create_experiment_experiments_post",
    "parameters": [
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExperimentRequest"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/experiments/{experiment_id}": {
   "get": {
    "summary": "Get Experiment",
    "operationId": "get_experiment_experiments__experiment_id__get",
    "parameters": [
     {
      "name": "experiment_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Experiment Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Experiment Experiments  Experiment Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "CreateSession": {
    "properties": {
     "employee_id": {
      "type": "string",
      "title": "Employee Id"
     },
     "config_ref": {
      "type": "string",
      "title": "Config Ref",
      "default": "agent_config"
     },
     "metadata": {
      "additionalProperties": true,
      "type": "object",
      "title": "Metadata"
     }
    },
    "type": "object",
    "required": [
     "employee_id"
    ],
    "title": "CreateSession"
   },
   "ExperimentRequest": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "questions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Questions"
     },
     "basket_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Basket Id"
     },
     "employee_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Employee Ids"
     },
     "config_matrix": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Config Matrix"
     },
     "max_tokens": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Tokens"
     },
     "max_usd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Usd"
     },
     "expected_calls_per_question": {
      "type": "integer",
      "title": "Expected Calls Per Question",
      "default": 6
     }
    },
    "type": "object",
    "required": [
     "name",
     "employee_ids"
    ],
    "title": "ExperimentRequest",
    "description": "A batch over a config matrix.\n\n``budget`` is required. An experiment without a declared ceiling is refused\nrather than defaulted, because a default ceiling is one nobody chose."
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}